LayoutTests/editing/smart-lists/backspace-immediately-after-smart-list-creation-undoes-list-expected.txt

Pressing Backspace immediately after Smart Lists converts two consecutive lines into a list should undo the conversion, restoring the original markers and the trigger space.
| <div>
|   "* A"
| <div>
|   "* "
|   <#selection-caret>
